refactor update and create validator modals into one
