+++ frontend/src/components/Simulator/ValidatorItem.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { type ValidatorModel } from '@/types'
 import { PencilSquareIcon, TrashIcon } from '@heroicons/vue/24/solid'
-import UpdateValidatorModal from '@/components/Simulator/UpdateValidatorModal.vue'
+import ValidatorModal from '@/components/Simulator/ValidatorModal.vue'
 import DeleteValidatorModal from '@/components/Simulator/DeleteValidatorModal.vue'
 import { ref } from 'vue'
 
@@ -52,7 +52,7 @@ const props = defineProps<{
     </div>
   </div>
 
-  <UpdateValidatorModal
+  <ValidatorModal
     :validator="validator"
     :open="isUpdateModalMopen"
     @close="isUpdateModalMopen = false"
